refactor: Round 3 — split game_routes, extract lifecycle constants and goals
game_routes.py (1054 → 26 line aggregator):
  - Split into 7 focused route modules under api_gateway/routes/:
    auth_routes.py (66), world_routes.py (279), federation_routes.py (112),
    wrestler_routes.py (126), booking_routes.py (202),
    storyline_routes.py (132), stable_manager_routes.py
  - game_routes.py is now a lightweight router aggregator

wrestler_lifecycle_service.py (in progress):
  - Extract lifecycle_constants.py (214 lines) — retirement thresholds,
    goal types, training modifiers, conditioning formulas
  - Extract goal_service.py — goal creation and evaluation logic

character_agent.py refactoring still in progress.
